% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_recommended_dose}
\alias{get_recommended_dose}
\title{get_recommended_dose() Function}
\usage{
get_recommended_dose(
  mode,
  model = NULL,
  target = NULL,
  current_dose = NULL,
  current_status = NULL,
  ...
)
}
\description{
This function adds a and b
}
\examples{
get_recommended_dose()
}
\keyword{test}
