% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmctdmr.R
\name{hmctdmrest}
\alias{hmctdmrest}
\title{hmctdmr() Function}
\usage{
hmctdmrest(
  drug = NULL,
  data = NULL,
  prior = NULL,
  stan_model_option = NULL,
  stan_sample_option = NULL
)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function adds a and b
}
\examples{
hmctdmr()
}
\keyword{hmctdmr}
